Setup unit testing
