+++ Assets/Citadel.asmdef
@@ -1,0 +1,15 @@
+{
+    "name": "Citadel",
+    "references": [
+        "GUID:c685d05731421f64287ad6d13be5af0e"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
