+++ Assets/Citadel.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fa1486058d6b204ae84528a85f5efe89
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
