--- Assets/Scripts/Tests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db926b29e75b7c94b9a58a9a48206b57
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
