+++ Assets/Tests/TestGeneralSetup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb3b40e6c97feda2298c759fbed6c01e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
